split up template rule
