+++ src/rules/switchTemplateExportAsNamesRule.ts
@@ -1,0 +1,66 @@
+import {green, red} from 'chalk';
+import {Replacement, RuleFailure, Rules} from 'tslint';
+import * as ts from 'typescript';
+import {exportAsNames} from '../material/component-data';
+import {ExternalResource} from '../tslint/component-file';
+import {ComponentWalker} from '../tslint/component-walker';
+import {findAll} from '../typescript/literal';
+
+/**
+ * Rule that walks through every component decorator and updates their inline or external
+ * templates.
+ */
+export class Rule extends Rules.AbstractRule {
+  apply(sourceFile: ts.SourceFile): RuleFailure[] {
+    return this.applyWithWalker(
+        new SwitchTemplateExportAsNamesWalker(sourceFile, this.getOptions()));
+  }
+}
+
+export class SwitchTemplateExportAsNamesWalker extends ComponentWalker {
+  visitInlineTemplate(template: ts.StringLiteral) {
+    this.replaceNamesInTemplate(template, template.getText()).forEach(replacement => {
+      const fix = replacement.replacement;
+      const ruleFailure = new RuleFailure(template.getSourceFile(), fix.start, fix.end,
+          replacement.message, this.getRuleName(), fix);
+      this.addFailure(ruleFailure);
+    });
+  }
+
+  visitExternalTemplate(template: ExternalResource) {
+    this.replaceNamesInTemplate(template, template.getFullText()).forEach(replacement => {
+      const fix = replacement.replacement;
+      const ruleFailure = new RuleFailure(template, fix.start + 1, fix.end + 1,
+          replacement.message, this.getRuleName(), fix);
+      this.addFailure(ruleFailure);
+    });
+  }
+
+  /**
+   * Replaces the outdated name in the template with the new one and returns an updated template.
+   */
+  private replaceNamesInTemplate(node: ts.Node, templateContent: string):
+      {message: string, replacement: Replacement}[] {
+    const replacements: {message: string, replacement: Replacement}[] = [];
+
+    exportAsNames.forEach(name => {
+      this.createReplacementsForOffsets(node, name, findAll(templateContent, name.replace))
+          .forEach(replacement => {
+            replacements.push({
+              message: `Found deprecated exportAs reference "${red(name.replace)}" which has been` +
+              ` renamed to "${green(name.replaceWith)}"`,
+              replacement
+            });
+          })
+    });
+
+    return replacements;
+  }
+
+  private createReplacementsForOffsets(node: ts.Node,
+                                       update: {replace: string, replaceWith: string},
+                                       offsets: number[]): Replacement[] {
+    return offsets.map(offset => this.createReplacement(
+        node.getStart() + offset, update.replace.length, update.replaceWith));
+  }
+}
