+++ test/fixtures/sample-project/app/app.component.html
@@ -11,8 +11,8 @@
 <mat-select (change)="" stuff></mat-select>
 <div class="mat-input-container"></div>
 <mat-input-container></mat-input-container>
-<mat-basic-chip (destroyed)="">
-<div mat-basic-chip (destroyed)=""></div>
+<mat-basic-chip (destroy)="">
+<div mat-basic-chip (destroy)=""></div>
 <div mat-chip="stuff" (remove)=""></div>
 
 <!-- These lines should not be changed: -->
